[I2C] Re-organizing code to make future maintenance easier
Both ARC and non-ARC boards will now use the same base code.  The
IPM logic has been moved to its own file.  Now any changed needed
for I2C can address either the I2C code or the IPM code.  This
will remove the need to update two code bases with any future
changes.
